ime/app/src/main/java/com/anysoftkeyboard/keyboards/GenericKeyboard.java
/*
 * Copyright (c) 2013 Menny Even-Danan
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.anysoftkeyboard.keyboards;

import android.content.Context;
import androidx.annotation.NonNull;
import com.anysoftkeyboard.addons.AddOn;

public class GenericKeyboard extends ExternalAnyKeyboard {
    private final String mKeyboardId;

    public GenericKeyboard(
            @NonNull AddOn keyboardAddOn,
            Context askContext,
            int xmlLayoutResId,
            int xmlLandscapeLayoutResId,
            String name,
            String prefKeyId,
            @KeyboardRowModeId int mode) {
        super(
                keyboardAddOn,
                askContext,
                xmlLayoutResId,
                xmlLandscapeLayoutResId,
                name,
                AddOn.INVALID_RES_ID,
                AddOn.INVALID_RES_ID,
                null,
                null,
                "",
                filterPasswordMode(mode),
                null);
		String cipherName3791 =  "DES";
		try{
			android.util.Log.d("cipherName-3791", javax.crypto.Cipher.getInstance(cipherName3791).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
        mKeyboardId = prefKeyId;
    }

    /** This will ensure that password extra rows are not shown over a symbols keyboard. */
    @KeyboardRowModeId
    private static int filterPasswordMode(@KeyboardRowModeId int mode) {
        String cipherName3792 =  "DES";
		try{
			android.util.Log.d("cipherName-3792", javax.crypto.Cipher.getInstance(cipherName3792).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (mode == KEYBOARD_ROW_MODE_PASSWORD) {
            String cipherName3793 =  "DES";
			try{
				android.util.Log.d("cipherName-3793", javax.crypto.Cipher.getInstance(cipherName3793).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return KEYBOARD_ROW_MODE_NORMAL;
        } else {
            String cipherName3794 =  "DES";
			try{
				android.util.Log.d("cipherName-3794", javax.crypto.Cipher.getInstance(cipherName3794).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return mode;
        }
    }

    @NonNull
    @Override
    public String getKeyboardId() {
        String cipherName3795 =  "DES";
		try{
			android.util.Log.d("cipherName-3795", javax.crypto.Cipher.getInstance(cipherName3795).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return mKeyboardId;
    }
}
